slightly nicer validation of API tokens.
It's still a bit naff.
doCheckCredentialsId method get called multiple times (🤯) - twice on page load and 4 times when you change the credentialsId, so it's infeasible to call a CBFM API each time in doCheckCredentialsId().
Also, the CBFM public API is rate limited, so we can't hit it too many times.

Still, this is better than just throwing an angry-jenkins error.